backend/requirements.txt
Flask
python-dotenv
yfinance
requests
pandas
numpy
schedule
feedparser
psycopg2-binary
redis